chore: add validation to delete question use case

+++ src/domain/forum/application/use-cases/delete-question-use-case.ts
@@ -2,6 +2,7 @@ import { QuestionsRepository } from '../repositories/questions-repository';
 
 interface DeleteQuestionRequest {
 	questionId: string;
+	authorId: string;
 }
 
 interface DeleteQuestionResponse {}
@@ -11,11 +12,16 @@ export class DeleteQuestionUseCase {
 
 	async execute({
 		questionId,
+		authorId,
 	}: DeleteQuestionRequest): Promise<DeleteQuestionResponse> {
 		const question = await this.questionsRepository.findById(questionId);
 
 		if (!question) {
 			throw new Error('Question not found.');
+		}
+
+		if (authorId !== question.authorId.toString()) {
+			throw new Error('Not allowed');
 		}
 
 		await this.questionsRepository.delete(question);
